Refactor syringe calibration logic
