+++ syringe-filler-pio/src/app/SyringeFillController.cpp
@@ -1,7 +1,6 @@
 #include "app/SyringeFillController.hpp"
 
 #include "hw/Bases.hpp"
-#include "hw/Pots.hpp"
 #include "motion/Axis.hpp"
 #include "hw/RFID.hpp"
 #include "hw/BaseRFID.hpp"
@@ -13,7 +12,6 @@ namespace App {
 
 namespace {
   constexpr bool SFC_DBG = true;
-  constexpr uint8_t TOOLHEAD_POT_IDX = 0;  // TODO: set your toolhead pot index
 
   // used by BaseRFID listener
   struct BaseTagCapture {
@@ -41,23 +39,13 @@ namespace {
     }
   }
 
-  void syncLegacyFields(PotCalibration& cal) {
-    if (cal.pointCount == 0) return;
-    cal.adcEmpty = Pots::countsFromRatio(cal.points[0].ratio);
-    if (cal.pointCount > 1) {
-      cal.adcFull = Pots::countsFromRatio(cal.points[cal.pointCount - 1].ratio);
-      cal.mlFull  = cal.points[cal.pointCount - 1].volume_ml;
-    } else {
-      cal.adcFull = cal.adcEmpty;
-      cal.mlFull  = cal.points[0].volume_ml;
-    }
-  }
 }
 
 // ------------------------------------------------------------
 // ctor
 // ------------------------------------------------------------
-SyringeFillController::SyringeFillController() {
+SyringeFillController::SyringeFillController()
+  : m_calibration(m_toolhead, m_bases, Bases::kCount, m_baseToPot, m_currentSlot) {
   if (SFC_DBG) {
     Serial.println("[SFC] ctor: init base slots");
   }
@@ -75,39 +63,8 @@ SyringeFillController::SyringeFillController() {
 }
 
 
-int8_t SyringeFillController::getBasePotIndex(uint8_t baseSlot) const {
-  if (baseSlot >= Bases::kCount) return -1;
-  uint8_t p = m_baseToPot[baseSlot];
-  return (p == 0xFF) ? -1 : (int8_t)p;
-}
-
-// ------------------------------------------------------------
-// toolhead calibration helpers
-// ------------------------------------------------------------
-float SyringeFillController::readToolheadRatio() {
-  Pots::poll();
-  float ratio = Pots::percent(TOOLHEAD_POT_IDX);
-  if (SFC_DBG) {
-    Serial.print("[SFC] readToolheadRatio(): pot=");
-    Serial.print(TOOLHEAD_POT_IDX);
-    Serial.print(" ratio=");
-    Serial.println(ratio, 3);
-  }
-  return ratio;
-}
-
 bool SyringeFillController::saveToolheadCalibration() {
-  if (m_toolhead.rfid == 0) {
-    if (SFC_DBG) Serial.println("[SFC] saveToolheadCalibration(): no toolhead RFID");
-    return false;
-  }
-  syncLegacyFields(m_toolhead.cal);
-  bool ok = Util::saveCalibration(m_toolhead.rfid, m_toolhead.cal);
-  if (SFC_DBG) {
-    Serial.print("[SFC] saveToolheadCalibration(): ");
-    Serial.println(ok ? "OK" : "FAIL");
-  }
-  return ok;
+  return m_calibration.saveToolheadCalibration();
 }
 
 // ------------------------------------------------------------
@@ -165,36 +122,10 @@ bool SyringeFillController::scanBaseSyringe(uint8_t slot) {
     Serial.println(tag, HEX);
   }
 
-  // update runtime object
+  m_calibration.initializeBaseFromTag(slot, tag);
+
   Syringe& s = m_bases[slot];
-  s.rfid = tag;
-  s.role = SyringeRole::Base;
-  s.slot = slot;
-// try to load from NVS
-Util::BaseMeta meta;
-PotCalibration cal;
-App::CalibrationPoints points;
-if (Util::loadBase(tag, meta, cal, points)) {
-  s.cal = cal;
-  s.calPoints = points;
-  if (SFC_DBG) {
-    Serial.print("[SFC] scanBaseSyringe(): existing base loaded from NVS for tag 0x");
-    Serial.println(tag, HEX);
-  }
-} else {
-  // NEW BASE: create minimal meta and save immediately (no slot stored)
-  if (SFC_DBG) {
-    Serial.print("[SFC] scanBaseSyringe(): NEW base, creating NVS entry for tag 0x");
-    Serial.println(tag, HEX);
-  }
-  memset(&meta, 0, sizeof(meta));
-  s.calPoints.count = 0;
-  Util::saveBase(tag, meta, s.cal, s.calPoints);
-}
-
-
-  // also update current volume (stub for now)
-  s.currentMl = readBaseVolumeMl(slot);
+  s.currentMl = m_calibration.readBaseVolumeMl(slot);
   if (SFC_DBG) {
     Serial.print("[SFC] scanBaseSyringe(): measured ~");
     Serial.print(s.currentMl, 3);
@@ -204,359 +135,34 @@ if (Util::loadBase(tag, meta, cal, points)) {
   return true;
 }
 bool SyringeFillController::captureToolheadCalibrationPoint(float ml, String& message) {
-  if (ml < 0.0f) {
-    message = "volume must be >= 0";
-    return false;
-  }
-  if (m_toolhead.rfid == 0) {
-    message = "no toolhead RFID; run sfc.scanTool first";
-    return false;
-  }
-
-  float ratio = readToolheadRatio();
-  bool ok = m_toolhead.cal.addPoint(ml, ratio);
-  if (!ok) {
-    message = (m_toolhead.cal.pointCount >= PotCalibration::kMaxPoints)
-                ? "toolhead calibration point list full"
-                : "failed to add toolhead calibration point";
-    return false;
-  }
-
-  syncLegacyFields(m_toolhead.cal);
-  if (m_toolhead.cal.pointCount < 2) {
-    message = "point saved; add at least 2 points to enable interpolation";
-  } else {
-    message = "point saved";
-  }
-  return true;
+  return m_calibration.captureToolheadCalibrationPoint(ml, message);
 }
 
 bool SyringeFillController::clearCurrentBaseCalibrationPoints(String& message) {
-  if (m_currentSlot < 0 || m_currentSlot >= (int)Bases::kCount) {
-    message = "no current base";
-    return false;
-  }
-
-  Syringe& sy = m_bases[m_currentSlot];
-  if (sy.rfid == 0) {
-    message = "base has no RFID; run sfc.scanbase first";
-    return false;
-  }
-
-  sy.calPoints = {};
-  Util::BaseMeta meta;
-  App::PotCalibration cal;
-  App::CalibrationPoints savedPoints;
-  if (!Util::loadBase(sy.rfid, meta, cal, savedPoints)) {
-    memset(&meta, 0, sizeof(meta));
-  }
-  (void)cal;
-  (void)savedPoints;
-  bool ok = Util::saveBase(sy.rfid, meta, sy.cal, sy.calPoints);
-  message = ok ? "base calibration points cleared" : "failed to clear base calibration points";
-  return ok;
+  return m_calibration.clearCurrentBaseCalibrationPoints(message);
 }
 
 bool SyringeFillController::clearToolheadCalibrationPoints(String& message) {
-  if (m_toolhead.rfid == 0) {
-    message = "no toolhead RFID; run sfc.scanTool first";
-    return false;
-  }
-
-  m_toolhead.cal.pointCount = 0;
-  m_toolhead.cal.adcEmpty = 0;
-  m_toolhead.cal.adcFull = 0;
-  m_toolhead.cal.mlFull = 0.0f;
-  bool ok = Util::saveCalibration(m_toolhead.rfid, m_toolhead.cal);
-  message = ok ? "toolhead calibration points cleared" : "failed to clear toolhead calibration points";
-  return ok;
+  return m_calibration.clearToolheadCalibrationPoints(message);
 }
 
 bool SyringeFillController::forceCurrentBaseCalibrationZero(String& message) {
-  if (m_currentSlot < 0 || m_currentSlot >= (int)Bases::kCount) {
-    message = "no current base";
-    return false;
-  }
-
-  Syringe& sy = m_bases[m_currentSlot];
-  if (sy.rfid == 0) {
-    message = "base has no RFID; run sfc.scanbase first";
-    return false;
-  }
-  if (sy.calPoints.count == 0) {
-    message = "no base calibration points to update";
-    return false;
-  }
-
-  sy.calPoints.points[0].volumeMl = 0.0f;
-  Util::BaseMeta meta;
-  App::PotCalibration cal;
-  App::CalibrationPoints savedPoints;
-  if (!Util::loadBase(sy.rfid, meta, cal, savedPoints)) {
-    memset(&meta, 0, sizeof(meta));
-  }
-  (void)cal;
-  (void)savedPoints;
-  bool ok = Util::saveBase(sy.rfid, meta, sy.cal, sy.calPoints);
-  message = ok ? "base calibration forced through 0 mL" : "failed to update base calibration";
-  return ok;
+  return m_calibration.forceCurrentBaseCalibrationZero(message);
 }
 
 bool SyringeFillController::forceToolheadCalibrationZero(String& message) {
-  if (m_toolhead.rfid == 0) {
-    message = "no toolhead RFID; run sfc.scanTool first";
-    return false;
-  }
-  if (m_toolhead.cal.pointCount == 0) {
-    message = "no toolhead calibration points to update";
-    return false;
-  }
-
-  m_toolhead.cal.points[0].volume_ml = 0.0f;
-  syncLegacyFields(m_toolhead.cal);
-  bool ok = Util::saveCalibration(m_toolhead.rfid, m_toolhead.cal);
-  message = ok ? "toolhead calibration forced through 0 mL" : "failed to update toolhead calibration";
-  return ok;
+  return m_calibration.forceToolheadCalibrationZero(message);
 }
-
-
-float SyringeFillController::readBaseRatio(uint8_t slot) {
-  if (slot >= Bases::kCount) {
-    if (SFC_DBG) {
-      Serial.print("[SFC] readBaseRatio(): slot OOR ");
-      Serial.println(slot);
-    }
-    return 0.0f;
-  }
-
-  int8_t potIdx = getBasePotIndex(slot);
-  if (potIdx < 0) {
-    if (SFC_DBG) {
-      Serial.print("[SFC] readBaseRatio(): no pot mapped for base ");
-      Serial.println(slot);
-    }
-    return 0.0f;
-  }
-
-  Pots::poll();
-  float ratio = Pots::percent((uint8_t)potIdx);
-
-  if (SFC_DBG) {
-    Serial.print("[SFC] readBaseRatio(base=");
-    Serial.print(slot);
-    Serial.print(" pot=");
-    Serial.print(potIdx);
-    Serial.print("): ratio=");
-    Serial.print(ratio, 3);
-  }
-
-  return ratio;
-}
-
-bool SyringeFillController::readBasePotRatio(uint8_t slot, float& ratio, String& message) const {
-  if (slot >= Bases::kCount) {
-    message = "base slot out of range";
-    return false;
-  }
-
-  int8_t potIdx = getBasePotIndex(slot);
-  if (potIdx < 0) {
-    message = "no pot mapped for base";
-    return false;
-  }
-  
-  Pots::poll();
-  float percent = Pots::percent((uint8_t)potIdx);
-  ratio = percent / 100.0f;
-  if (ratio < 0.0f) ratio = 0.0f;
-  if (ratio > 1.0f) ratio = 1.0f;
-  return true;
-}
-
-float SyringeFillController::interpolateVolumeFromPoints(const App::CalibrationPoints& points, float ratio, bool& ok) {
-  ok = false;
-  if (points.count < 2) return NAN;
-
-  if (ratio <= points.points[0].ratio) {
-    ok = true;
-    return points.points[0].volumeMl;
-  }
-  if (ratio >= points.points[points.count - 1].ratio) {
-    ok = true;
-    return points.points[points.count - 1].volumeMl;
-  }
-
-  for (uint8_t i = 0; i + 1 < points.count; ++i) {
-    const CalibrationPoint& a = points.points[i];
-    const CalibrationPoint& b = points.points[i + 1];
-    if (ratio >= a.ratio && ratio <= b.ratio) {
-      float denom = b.ratio - a.ratio;
-      if (denom <= 0.0f) return NAN;
-      float t = (ratio - a.ratio) / denom;
-      ok = true;
-      return a.volumeMl + t * (b.volumeMl - a.volumeMl);
-    }
-  }
-  return NAN;
-}
-
-// --------------------------------------------------
-// capture base EMPTY (pot @ fully empty)
-// --------------------------------------------------
-
 bool SyringeFillController::captureBaseCalibrationPoint(uint8_t slot, float ml, String& message) {
-  if (slot >= Bases::kCount) {
-    message = "base slot out of range";
-    return false;
-  }
-  if (ml < 0.0f) {
-    message = "volume must be >= 0";
-    return false;
-  }
-
-  Syringe& sy = m_bases[slot];
-  if (sy.rfid == 0) {
-    message = "base has no RFID; run sfc.scanbase first";
-    return false;
-  }
-
-  float ratio = 0.0f;
-  String ratioMessage;
-  if (!readBasePotRatio(slot, ratio, ratioMessage)) {
-    message = ratioMessage;
-    return false;
-  }
-
-  CalibrationPoints& points = sy.calPoints;
-  const float kRatioEpsilon = 0.0025f;
-  for (uint8_t i = 0; i < points.count; ++i) {
-    if (fabsf(points.points[i].ratio - ratio) <= kRatioEpsilon) {
-      points.points[i].volumeMl = ml;
-      Util::BaseMeta meta;
-      App::PotCalibration cal;
-      App::CalibrationPoints savedPoints;
-      if (!Util::loadBase(sy.rfid, meta, cal, savedPoints)) {
-        memset(&meta, 0, sizeof(meta));
-      }
-      (void)cal;
-      (void)savedPoints;
-      bool ok = Util::saveBase(sy.rfid, meta, sy.cal, sy.calPoints);
-      message = ok ? "updated calibration point" : "failed to save calibration point";
-      return ok;
-    }
-  }
-
-  if (points.count >= CalibrationPoints::kMaxPoints) {
-    message = "calibration point list full";
-    return false;
-  }
-
-  uint8_t insertAt = 0;
-  while (insertAt < points.count && points.points[insertAt].ratio < ratio) {
-    insertAt++;
-  }
-  for (uint8_t i = points.count; i > insertAt; --i) {
-    points.points[i] = points.points[i - 1];
-  }
-  points.points[insertAt].volumeMl = ml;
-  points.points[insertAt].ratio = ratio;
-  points.count++;
-
-  Util::BaseMeta meta;
-  App::PotCalibration cal;
-  App::CalibrationPoints savedPoints;
-  if (!Util::loadBase(sy.rfid, meta, cal, savedPoints)) {
-    memset(&meta, 0, sizeof(meta));
-  }
-  (void)cal;
-  (void)savedPoints;
-  bool ok = Util::saveBase(sy.rfid, meta, sy.cal, sy.calPoints);
-  if (!ok) {
-    message = "failed to save calibration point";
-    return false;
-  }
-
-  if (points.count < 2) {
-    message = "point saved; add at least 2 points to enable interpolation";
-  } else {
-    message = "point saved";
-  }
-  return true;
+  return m_calibration.captureBaseCalibrationPoint(slot, ml, message);
 }
 
-// --------------------------------------------------
-// save current base (whatever slot we are at) to NVS
-// --------------------------------------------------
 bool SyringeFillController::saveCurrentBaseToNVS() {
-  if (m_currentSlot < 0 || m_currentSlot >= (int8_t)Bases::kCount) {
-    if (SFC_DBG) Serial.println("[SFC] saveCurrentBaseToNVS(): no current slot");
-    return false;
-  }
-
-  Syringe& sy = m_bases[m_currentSlot];
-  if (sy.rfid == 0) {
-    if (SFC_DBG) Serial.println("[SFC] saveCurrentBaseToNVS(): no RFID cached for this slot");
-    return false;
-  }
-
-  Util::BaseMeta meta;
-  PotCalibration cal;
-  App::CalibrationPoints points;
-  if (Util::loadBase(sy.rfid, meta, cal, points)) {
-    // update cal
-    cal = sy.cal;
-    if (SFC_DBG) {
-      Serial.print("[SFC] saveCurrentBaseToNVS(): updating existing base 0x");
-      Serial.println(sy.rfid, HEX);
-    }
-    (void)points;
-    return Util::saveBase(sy.rfid, meta, cal, sy.calPoints);
-  } else {
-    // create new
-    if (SFC_DBG) {
-      Serial.print("[SFC] saveCurrentBaseToNVS(): creating new base 0x");
-      Serial.println(sy.rfid, HEX);
-    }
-    memset(&meta, 0, sizeof(meta));
-    return Util::saveBase(sy.rfid, meta, sy.cal, sy.calPoints);
-  }
+  return m_calibration.saveCurrentBaseToNVS();
 }
 
-// --------------------------------------------------
-// print helper
-// --------------------------------------------------
 void SyringeFillController::printBaseInfo(uint8_t slot, Stream& s) {
-  if (slot >= Bases::kCount) {
-    s.println("[SFC] base info: slot OOR");
-    return;
-  }
-  const Syringe& sy = m_bases[slot];
-  s.print("[SFC] base "); s.print(slot);
-  s.print(" RFID=0x"); s.print(sy.rfid, HEX);
-  s.print(" points="); s.print(sy.cal.pointCount);
-  for (uint8_t i = 0; i < sy.cal.pointCount; ++i) {
-    s.print(" [");
-    s.print(i);
-    s.print(":");
-    s.print(sy.cal.points[i].volume_ml, 3);
-    s.print("ml@");
-    s.print(sy.cal.points[i].ratio, 3);
-    s.print("]");
-  }
-  s.print(" steps_mL=");   s.print(sy.cal.steps_mL, 3);
-  s.print(" calPoints="); s.print(sy.calPoints.count);
-  if (sy.calPoints.count > 0) {
-    s.print(" [");
-    for (uint8_t i = 0; i < sy.calPoints.count; ++i) {
-      if (i > 0) s.print(", ");
-      s.print(sy.calPoints.points[i].volumeMl, 3);
-      s.print("ml@");
-      s.print(sy.calPoints.points[i].ratio * 100.0f, 2);
-      s.print("%");
-    }
-    s.print("]");
-  }
-  s.println();
+  m_calibration.printBaseInfo(slot, s);
 }
 // ------------------------------------------------------------
 // blocking base-RFID read (with tick)
@@ -641,20 +247,9 @@ bool SyringeFillController::scanToolheadBlocking() {
     return false;
   }
 
-  m_toolhead.rfid = tag;
-  m_toolhead.role = SyringeRole::Toolhead;
+  m_calibration.initializeToolheadFromTag(tag);
 
-  PotCalibration cal;
-  if (Util::loadCalibration(tag, cal)) {
-    m_toolhead.cal = cal;
-    Serial.print("[SFC] loaded toolhead cal for 0x");
-    Serial.println(tag, HEX);
-  } else {
-    Serial.print("[SFC] no toolhead cal for 0x");
-    Serial.println(tag, HEX);
-  }
-
-  m_toolhead.currentMl = readToolheadVolumeMl();
+  m_toolhead.currentMl = m_calibration.readToolheadVolumeMl();
   Serial.print("[SFC] toolhead volume ~");
   Serial.println(m_toolhead.currentMl, 3);
 
@@ -728,57 +323,8 @@ uint32_t SyringeFillController::readToolheadRFIDBlocking(uint32_t timeoutMs) {
   return cap.packed;
 }
 
-float App::SyringeFillController::mlFromRatio_(const App::PotCalibration& cal, float ratio) {
-  return cal.ratioToMl(ratio);
-}
-
 void App::SyringeFillController::printToolheadInfo(Stream& out) {
-  out.println(F("[SFC] Toolhead calibration:"));
-
-  if (m_toolhead.rfid == 0) {
-    out.println(F("  RFID: (none)  tip: run 'sfc.scanTool'"));
-    return;
-  }
-
-  out.print(F("  RFID: 0x"));
-  out.println(m_toolhead.rfid, HEX);
-
-  // Refresh calibration from NVS each time (so it always matches Storage.cpp)
-  App::PotCalibration calTmp;
-  bool ok = Util::loadCalibration(m_toolhead.rfid, calTmp);
-  m_toolCalValid = ok;
-  if (ok) m_toolCal = calTmp;
-
-  out.print(F("  cal loaded from memory: "));
-  out.println(ok ? F("yes") : F("no"));
-
-  if (ok) {
-    out.print(F("  cal.points(")); out.print(m_toolCal.pointCount); out.println(F("):"));
-    for (uint8_t i = 0; i < m_toolCal.pointCount; ++i) {
-      out.print(F("    ["));
-      out.print(i);
-      out.print(F("] "));
-      out.print(m_toolCal.points[i].volume_ml, 3);
-      out.print(F(" ml @ "));
-      out.print(m_toolCal.points[i].ratio, 3);
-      out.println(F(" ratio"));
-    }
-    out.print(F("  cal.steps_mL     : ")); out.println(m_toolCal.steps_mL, 3);
-  }
-
-  // Live pot read (set this index correctly for the toolhead pot)
-  float ratio = Pots::percent(TOOLHEAD_POT_IDX);
-  uint16_t scaled = Pots::readScaled(TOOLHEAD_POT_IDX);
-
-  out.print(F("  pot.ratio: ")); out.println(ratio, 3);
-  out.print(F("  pot.scaled: ")); out.println(scaled);
-
-  if (ok) {
-    float ml = mlFromRatio_(m_toolCal, ratio);
-    out.print(F("  computed mL: "));
-    if (isnan(ml)) out.println(F("(n/a)"));
-    else out.println(ml, 3);
-  }
+  m_calibration.printToolheadInfo(out);
 }
 
 
@@ -830,7 +376,7 @@ void SyringeFillController::runRecipe() {
       }
     }
   }
-  m_toolhead.currentMl = readToolheadVolumeMl();
+  m_toolhead.currentMl = m_calibration.readToolheadVolumeMl();
   dbg("runRecipe() done");
 }
 
@@ -919,54 +465,6 @@ uint32_t SyringeFillController::readRFIDNow() {
   if (SFC_DBG) Serial.println("[SFC] readRFIDNow(): timeout, no tag");
   if (!wasEnabled) RFID::enable(false);
   return 0;
-}
-
-// ------------------------------------------------------------
-// volume helpers (still stubs)
-// ------------------------------------------------------------
-float SyringeFillController::readToolheadVolumeMl() {
-  float ratio = readToolheadRatio();
-  float ml = m_toolhead.cal.ratioToMl(ratio);
-  if (SFC_DBG) {
-    Serial.print("[SFC] readToolheadVolumeMl(): stub -> ");
-    Serial.println(ml, 3);
-  }
-  return ml;
-}
-
-float SyringeFillController::readBaseVolumeMl(uint8_t slot) {
-  if (slot >= Bases::kCount) {
-    if (SFC_DBG) {
-      Serial.print("[SFC] readBaseVolumeMl(): slot OOR ");
-      Serial.println(slot);
-    }
-    return 0.0f;
-  }
-
-  float ratio = 0.0f;
-  String ratioMessage;
-  bool ratioOk = readBasePotRatio(slot, ratio, ratioMessage);
-  if (ratioOk && m_bases[slot].calPoints.count >= 2) {
-    bool ok = false;
-    float ml = interpolateVolumeFromPoints(m_bases[slot].calPoints, ratio, ok);
-    if (ok) {
-      if (SFC_DBG) {
-        Serial.print("[SFC] readBaseVolumeMl(slot=");
-        Serial.print(slot);
-        Serial.print("): ratio=");
-        Serial.print(ratio, 4);
-        Serial.print(" -> ");
-        Serial.println(ml, 3);
-      }
-      return ml;
-    }
-  } else if (SFC_DBG && ratioOk) {
-    Serial.print("[SFC] readBaseVolumeMl(slot=");
-    Serial.print(slot);
-    Serial.println("): insufficient calibration points for interpolation");
-    return 0.0f;
-  }
-
 }
 
 // ------------------------------------------------------------
